Upgrade solution projects to target .NET 7
This commit upgrades the target framework of all solution projects from .NET 5 to .NET 7. The upgrade was made to take advantage of the latest features and performance improvements in .NET 7, and to ensure that the solution remains up-to-date and compatible with the latest versions of .NET.

The upgrade was made by editing the TargetFramework property in the .csproj files for each project and then resolving any build errors or warnings that arose as a result of the upgrade. In most cases, the changes were minor and involved updating references to packages or changing the syntax in a few places.

The upgrade has been tested and verified to work correctly. Overall, this upgrade should help to ensure that the solution is more stable, performant, and maintainable in the long term.