Implement Domain Entities and Customer Validation
In this commit, we have implemented the Customer entity with its properties in the Domain project. The Customer entity has properties for FirstName, LastName, DateOfBirth, PhoneNumber, Email, and BankAccountNumber. Additionally, we have defined a FluentValidation CustomerValidator in the same namespace to handle validations on the Customer entity.

We have added DataAnnotation attributes for some of the validations, and we have added custom validation for the uniqueness of the combination of FirstName, LastName, and DateOfBirth.

We have also defined the ICustomerRepository interface, which will be used to interact with the data store for storing and retrieving Customer entities. The ICustomerRepository interface is located in the Mc2.CrudTest.Domain.Repositories namespace.