Implement CustomerDTO, CustomerService, ICustomerRepository and Register them in DI Container
This commit implements the CustomerDTO for transferring data between layers, ICustomerRepository interface for defining the repository contract and the CustomerService for implementing the application layer use cases and business logic. Also, the CustomerValidator was added to validate the customer input data. Furthermore, the CustomerService and CustomerValidator were registered in the DI container in the Startup class of the Presentation.Server project.